+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## HEAD
 
+* Remove data.ValidateInterface...; does not work as expected
 * Remove structure.Validating; does not work as expected
 * Refactor data normalizer to use common structure normalizer
 * Remove normalization from physical activity duration as necessary